fix: modify pipeline listener files and add utils
