+++ public/app/listeners/executeAlgorithm.js
@@ -1,13 +1,10 @@
-const fs = require('fs');
 const path = require('path');
 const logger = require('../logger');
 const { DATASETS_FOLDER } = require('../config/paths');
-const { createNewDataset } = require('./loadDataset');
 const { buildExecuteAlgorithmChannel } = require('../../shared/channels');
 const {
   PROGRAMMING_LANGUAGES,
   EXECUTION_STATUSES,
-  DATASET_TYPES,
   DATASETS_COLLECTION,
   ALGORITHMS_COLLECTION,
   EXECUTIONS_COLLECTION,
@@ -17,39 +14,27 @@ const {
 const executePythonAlgorithm = require('./executePythonAlgorithm');
 const {
   ERROR_UNKNOWN_PROGRAMMING_LANGUAGE,
-  ERROR_EXECUTION_PROCESS,
   ERROR_GENERAL,
 } = require('../../shared/errors');
 const {
   EXECUTE_ALGORITHM_SUCCESS,
   EXECUTE_ALGORITHM_ERROR,
-  EXECUTE_ALGORITHM_STOP,
-  EXECUTE_ALGORITHM_UPDATE,
 } = require('../../shared/types');
-const { parseValidationResult } = require('../utils/validation');
-const { cancelExecutionObject } = require('./cancelExecution');
+const {
+  buildOnRunCallback,
+  buildOnStopCallback,
+  buildOnErrorCallback,
+  buildCleanCallback,
+  buildOnLogCallback,
+} = require('../utils/execution');
+const { cancelExecutionById } = require('./cancelExecution');
+const createNewResultDataset = require('../utils/result');
 const {
   ALGORITHM_DATASET_PATH_NAME,
-  ALGORITHM_OUTPUT_PATH_NAME,
   ALGORITHM_ORIGIN_PATH_NAME,
+  ALGORITHM_OUTPUT_PATH_NAME,
 } = require('../config/config');
-
-const createNewResultDataset = (
-  { name, filepath, algorithmId, description, originId },
-  db,
-) => {
-  const result = createNewDataset(
-    {
-      name,
-      filepath,
-      description,
-      type: DATASET_TYPES.RESULT,
-    },
-    db,
-  );
-
-  return { ...result, algorithmId, originId };
-};
+const { parseValidationResult } = require('../utils/validation');
 
 const executeAlgorithm = (mainWindow, db) => (event, { id: executionId }) => {
   const channel = buildExecuteAlgorithmChannel(executionId);
@@ -86,25 +71,7 @@ const executeAlgorithm = (mainWindow, db) => (event, { id: executionId }) => {
 
     // prepare onRun callback function
     // set execution as running and pid
-    const onRun = ({ pid }) => {
-      const execution = db
-        .get(EXECUTIONS_COLLECTION)
-        .find({ id: executionId })
-        .assign({ status: EXECUTION_STATUSES.RUNNING, pid })
-        .write();
-
-      // check whether mainWindow still exist in case of
-      // the app quits before the process get killed
-      return (
-        !mainWindow.isDestroyed() &&
-        mainWindow?.webContents?.send(channel, {
-          type: EXECUTE_ALGORITHM_UPDATE,
-          payload: {
-            execution,
-          },
-        })
-      );
-    };
+    const onRun = buildOnRunCallback(mainWindow, db, channel, { executionId });
 
     // prepare success callback function
     // copy tmp as new result dataset
@@ -169,73 +136,14 @@ const executeAlgorithm = (mainWindow, db) => (event, { id: executionId }) => {
       );
     };
 
-    // clean the tmp file at the end of the execution
-    const clean = () => {
-      if (fs.existsSync(tmpPath)) {
-        fs.unlinkSync(tmpPath);
-      }
-    };
-
-    const onStop = () => {
-      cancelExecutionObject(db, executionId);
-
-      // check whether mainWindow still exist in case of
-      // the app quits before the process get killed
-      return (
-        !mainWindow.isDestroyed() &&
-        mainWindow?.webContents?.send(channel, {
-          type: EXECUTE_ALGORITHM_STOP,
-        })
-      );
-    };
-
-    const onLog = ({ log }) => {
-      const execution = db
-        .get(EXECUTIONS_COLLECTION)
-        .find({ id: executionId })
-        .assign({ log })
-        .write();
-
-      // check whether mainWindow still exist in case of
-      // the app quits before the process get killed
-      return (
-        !mainWindow.isDestroyed() &&
-        mainWindow?.webContents?.send(channel, {
-          type: EXECUTE_ALGORITHM_UPDATE,
-          payload: {
-            execution,
-          },
-        })
-      );
-    };
-
-    // error handling when executing
-    const onError = ({ code, log }) => {
-      logger.error(
-        `process for execution ${executionId} exited with code ${code}`,
-      );
-      db.get(EXECUTIONS_COLLECTION)
-        .find({ id: executionId })
-        .assign({ status: EXECUTION_STATUSES.ERROR, log })
-        .unset('pid')
-        .write();
-
-      // check whether mainWindow still exist in case of
-      // the app quits before the process get killed
-      return (
-        !mainWindow.isDestroyed() &&
-        mainWindow?.webContents?.send(channel, {
-          type: EXECUTE_ALGORITHM_ERROR,
-          error: ERROR_EXECUTION_PROCESS,
-          payload: {
-            execution: db
-              .get(EXECUTIONS_COLLECTION)
-              .find({ id: executionId })
-              .value(),
-          },
-        })
-      );
-    };
+    const clean = buildCleanCallback(tmpPath);
+    const onStop = buildOnStopCallback(mainWindow, db, channel, {
+      executionId,
+    });
+    const onError = buildOnErrorCallback(mainWindow, db, channel, {
+      executionId,
+    });
+    const onLog = buildOnLogCallback(mainWindow, db, channel, { executionId });
 
     // path to the dataset
     const datasetPathParameter = {
@@ -301,4 +209,4 @@ const executeAlgorithm = (mainWindow, db) => (event, { id: executionId }) => {
   }
 };
 
-module.exports = executeAlgorithm;
+module.exports = { cancelExecutionById, executeAlgorithm };
