+++ public/app/listeners/executePipeline.js
@@ -1,32 +1,27 @@
 const logger = require('../logger');
 const { EXECUTE_PIPELINE_CHANNEL } = require('../../shared/channels');
-const { SAVE_PIPELINE_ERROR } = require('../../shared/types');
+const { EXECUTE_PIPELINE_ERROR } = require('../../shared/types');
 const { ERROR_GENERAL } = require('../../shared/errors');
-const runPipelineAlgorithm = require('./pipelineUtils.js');
+const executePipelineAlgorithm = require('./executePipelineAlgorithm.js');
 
 const executePipeline = (mainWindow, db) => (
   event,
-  { pipeline, sourceId, userProvidedFilename, parameters, schemaId },
+  { pipeline, sourceId, userProvidedFilename, schemaId },
 ) => {
   const { algorithms } = pipeline;
-  const algorithm = algorithms[0];
 
   try {
-    return runPipelineAlgorithm(
-      mainWindow,
-      db,
+    return executePipelineAlgorithm(mainWindow, db, {
       algorithms,
-      algorithm,
       sourceId,
       userProvidedFilename,
-      parameters,
       schemaId,
-      0,
-    );
+      algorithmIndex: 0,
+    });
   } catch (err) {
     logger.error(err);
     return mainWindow.webContents.send(EXECUTE_PIPELINE_CHANNEL, {
-      type: SAVE_PIPELINE_ERROR,
+      type: EXECUTE_PIPELINE_ERROR,
       error: ERROR_GENERAL,
     });
   }
