+++ public/app/listeners/index.js
@@ -1,5 +1,5 @@
 const { loadDataset } = require('./loadDataset');
-const executeAlgorithm = require('./executeAlgorithm');
+const { cancelExecutionById, executeAlgorithm } = require('./executeAlgorithm');
 const executePythonAlgorithm = require('./executePythonAlgorithm');
 const getDataset = require('./getDataset');
 const getDatasets = require('./getDatasets');
@@ -27,7 +27,7 @@ const getUtils = require('./getUtils');
 const saveUtils = require('./saveUtils');
 const getExecutions = require('./getExecutions');
 const deleteExecution = require('./deleteExecution');
-const { createExecution, createExecutionInDb } = require('./createExecution');
+const { createExecution, addExecutionObject } = require('./createExecution');
 const {
   cancelExecution,
   cancelExecutionObject,
@@ -93,8 +93,9 @@ module.exports = {
   getExecutions,
   deleteExecution,
   createExecution,
-  createExecutionInDb,
   cancelExecutionObject,
+  addExecutionObject,
+  cancelExecutionById,
   cancelAllRunningExecutions,
   cancelExecution,
   showConfirmDeletePrompt,
