+++ public/shared/types.js
@@ -150,6 +150,7 @@ const DELETE_VALIDATION_ERROR = 'DELETE_VALIDATION_ERROR';
 const EXECUTE_ALGORITHM_UPDATE = 'EXECUTE_ALGORITHM_UPDATE';
 const FLAG_EXECUTING_PIPELINE = 'FLAG_EXECUTING_PIPELINE';
 const EXECUTE_PIPELINE_SUCCESS = 'EXECUTE_PIPELINE_SUCCESS';
+const EXECUTE_PIPELINE_ERROR = 'EXECUTE_PIPELINE_ERROR';
 
 module.exports = {
   FLAG_GETTING_DATABASE,
@@ -301,4 +302,5 @@ module.exports = {
   EXECUTE_ALGORITHM_UPDATE,
   FLAG_EXECUTING_PIPELINE,
   EXECUTE_PIPELINE_SUCCESS,
+  EXECUTE_PIPELINE_ERROR,
 };
